Move logical expression type-coercion code from `physical-expr` crate to `expr` crate (#2257)
