+++ datafusion/core/src/execution/context.rs
@@ -1543,7 +1543,7 @@ mod tests {
     use super::*;
     use crate::execution::context::QueryPlanner;
     use crate::logical_plan::{binary_expr, create_udtf, lit, Operator};
-    use crate::physical_plan::functions::{make_table_function, Volatility};
+    use crate::physical_plan::functions::make_table_function;
     use crate::test;
     use crate::variable::VarType;
     use crate::{
@@ -1556,7 +1556,7 @@ mod tests {
     use arrow::datatypes::*;
     use arrow::record_batch::RecordBatch;
     use async_trait::async_trait;
-    use datafusion_expr::TableFunctionImplementation;
+    use datafusion_expr::{TableFunctionImplementation, Volatility};
     use std::fs::File;
     use std::sync::Weak;
     use std::thread::{self, JoinHandle};
