+++ datafusion/core/src/physical_plan/mod.rs
@@ -573,5 +573,4 @@ pub mod udf;
 pub mod udtf;
 pub mod union;
 pub mod values;
-pub mod window_functions;
 pub mod windows;
