+++ datafusion/expr/src/lib.rs
@@ -16,20 +16,25 @@
 // under the License.
 
 mod accumulator;
-mod aggregate_function;
+pub mod aggregate_function;
+pub mod array_expressions;
+pub mod binary_rule;
 mod built_in_function;
 mod columnar_value;
+pub mod conditional_expressions;
 pub mod expr;
 pub mod expr_fn;
-mod function;
+pub mod function;
 mod literal;
+mod nullif;
 mod operator;
 mod signature;
+pub mod type_coercion;
 mod udaf;
 mod udf;
 mod udtf;
-mod window_frame;
-mod window_function;
+pub mod window_frame;
+pub mod window_function;
 
 pub use accumulator::Accumulator;
 pub use aggregate_function::AggregateFunction;
@@ -42,6 +47,7 @@ pub use function::{
     StateTypeFunction, TableFunctionImplementation,
 };
 pub use literal::{lit, lit_timestamp_nano, Literal, TimestampLiteral};
+pub use nullif::SUPPORTED_NULLIF_TYPES;
 pub use operator::Operator;
 pub use signature::{Signature, TypeSignature, Volatility};
 pub use udaf::AggregateUDF;
