+++ datafusion/physical-expr/src/array_expressions.rs
@@ -105,22 +105,3 @@ pub fn array(values: &[ColumnarValue]) -> Result<ColumnarValue> {
 
     Ok(ColumnarValue::Array(array_array(&arrays)?))
 }
-
-/// Currently supported types by the array function.
-/// The order of these types correspond to the order on which coercion applies
-/// This should thus be from least informative to most informative
-pub static SUPPORTED_ARRAY_TYPES: &[DataType] = &[
-    DataType::Boolean,
-    DataType::UInt8,
-    DataType::UInt16,
-    DataType::UInt32,
-    DataType::UInt64,
-    DataType::Int8,
-    DataType::Int16,
-    DataType::Int32,
-    DataType::Int64,
-    DataType::Float32,
-    DataType::Float64,
-    DataType::Utf8,
-    DataType::LargeUtf8,
-];
