+++ datafusion/physical-expr/src/conditional_expressions.rs
@@ -20,7 +20,6 @@ use std::sync::Arc;
 use arrow::array::{new_null_array, Array, BooleanArray};
 use arrow::compute;
 use arrow::compute::kernels::zip::zip;
-use arrow::datatypes::DataType;
 
 use datafusion_common::{DataFusionError, Result};
 use datafusion_expr::ColumnarValue;
@@ -66,22 +65,3 @@ pub fn coalesce(args: &[ColumnarValue]) -> Result<ColumnarValue> {
 
     Ok(ColumnarValue::Array(Arc::new(res)))
 }
-
-/// Currently supported types by the coalesce function.
-/// The order of these types correspond to the order on which coercion applies
-/// This should thus be from least informative to most informative
-pub static SUPPORTED_COALESCE_TYPES: &[DataType] = &[
-    DataType::Boolean,
-    DataType::UInt8,
-    DataType::UInt16,
-    DataType::UInt32,
-    DataType::UInt64,
-    DataType::Int8,
-    DataType::Int16,
-    DataType::Int32,
-    DataType::Int64,
-    DataType::Float32,
-    DataType::Float64,
-    DataType::Utf8,
-    DataType::LargeUtf8,
-];
