+++ datafusion/physical-expr/src/expressions/covariance.rs
@@ -50,42 +50,6 @@ pub struct CovariancePop {
     expr2: Arc<dyn PhysicalExpr>,
 }
 
-/// function return type of covariance
-pub fn covariance_return_type(arg_type: &DataType) -> Result<DataType> {
-    match arg_type {
-        DataType::Int8
-        | DataType::Int16
-        | DataType::Int32
-        | DataType::Int64
-        | DataType::UInt8
-        | DataType::UInt16
-        | DataType::UInt32
-        | DataType::UInt64
-        | DataType::Float32
-        | DataType::Float64 => Ok(DataType::Float64),
-        other => Err(DataFusionError::Plan(format!(
-            "COVAR does not support {:?}",
-            other
-        ))),
-    }
-}
-
-pub fn is_covariance_support_arg_type(arg_type: &DataType) -> bool {
-    matches!(
-        arg_type,
-        DataType::UInt8
-            | DataType::UInt16
-            | DataType::UInt32
-            | DataType::UInt64
-            | DataType::Int8
-            | DataType::Int16
-            | DataType::Int32
-            | DataType::Int64
-            | DataType::Float32
-            | DataType::Float64
-    )
-}
-
 impl Covariance {
     /// Create a new COVAR aggregate function
     pub fn new(
@@ -521,17 +485,6 @@ mod tests {
             ScalarValue::from(0.6666666666666666_f64),
             DataType::Float64
         )
-    }
-
-    #[test]
-    fn test_covariance_return_data_type() -> Result<()> {
-        let data_type = DataType::Float64;
-        let result_type = covariance_return_type(&data_type)?;
-        assert_eq!(DataType::Float64, result_type);
-
-        let data_type = DataType::Decimal(36, 10);
-        assert!(covariance_return_type(&data_type).is_err());
-        Ok(())
     }
 
     #[test]
